+++ details.php
@@ -38,7 +38,7 @@ $link_results->execute();
   <script defer src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.12.2/gsap.min.js"></script>
   <script defer src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.12.2/MotionPathPlugin.min.js"></script>
   <script defer src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.12.2/ScrollTrigger.min.js"></script>
-  <script defer src="js/main.js"></script>
+  <script type="module" src="js/details.js"></script>
   <title><?php echo $row['title']; ?> | Isaac Bilyea</title>
 </head>
