+++ index.php
@@ -23,7 +23,7 @@ $stmt->execute();
   <script defer src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.12.2/gsap.min.js"></script>
   <script defer src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.12.2/MotionPathPlugin.min.js"></script>
   <script defer src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.12.2/ScrollTrigger.min.js"></script>
-  <script defer src="js/main.js"></script>
+  <script type="module" src="js/home.js"></script>
   <title>Home | Isaac Bilyea</title>
 </head>
