+++ js/about.js
@@ -1,0 +1,15 @@
+import { cursorDot } from './modules/cursor-dot.js';
+import { burgerMenu } from './modules/burger-menu.js';
+import { logoHover } from './modules/logo-hover.js';
+import { hueSlider } from './modules/hue-slider.js';
+import { imageSwitcher } from './modules/image-switcher.js';
+import { gsapAbout } from './modules/gsap-about.js';
+
+
+cursorDot();
+burgerMenu();
+logoHover();
+hueSlider();
+imageSwitcher();
+gsapAbout();
+
