+++ js/contact.js
@@ -1,0 +1,17 @@
+import { cursorDot } from './modules/cursor-dot.js';
+import { burgerMenu } from './modules/burger-menu.js';
+import { logoHover } from './modules/logo-hover.js';
+import { hueSlider } from './modules/hue-slider.js';
+import { contactBall } from './modules/contact-ball.js';
+import { ajaxForm } from './modules/ajax-form.js';
+import { gsapContact } from './modules/gsap-contact.js';
+
+
+cursorDot();
+burgerMenu();
+logoHover();
+hueSlider();
+contactBall();
+ajaxForm();
+gsapContact();
+
