+++ js/details.js
@@ -1,0 +1,15 @@
+import { cursorDot } from './modules/cursor-dot.js';
+import { burgerMenu } from './modules/burger-menu.js';
+import { logoHover } from './modules/logo-hover.js';
+import { hueSlider } from './modules/hue-slider.js';
+import { caseStudyScroll } from './modules/casestudy-scroll.js';
+import { gsapDetails } from './modules/gsap-details.js';
+
+
+cursorDot();
+burgerMenu();
+logoHover();
+hueSlider();
+caseStudyScroll();
+gsapDetails();
+
