+++ js/home.js
@@ -1,0 +1,16 @@
+import { cursorDot } from './modules/cursor-dot.js';
+import { burgerMenu } from './modules/burger-menu.js';
+import { logoHover } from './modules/logo-hover.js';
+import { hueSlider } from './modules/hue-slider.js';
+import { videoPlayer } from './modules/video-player.js';
+import { introBall } from './modules/intro-ball.js';
+import { gsapHome } from './modules/gsap-home.js';
+
+
+cursorDot();
+burgerMenu();
+logoHover();
+hueSlider();
+videoPlayer();
+introBall();
+gsapHome();
